requirements.txt
Flask==2.3.2
requests==2.31.0
google-cloud-storage==2.12.0
python-dotenv==1.0.0